Separated the jar files that need to get packaged inside of the jar file but not included in the .war file from the jars that are only needed for the build/test but should not be included in either package.
Changed the Log implementation in Service.java to use apache commons logging, instead of testng Logger.

Extended the delay for launching the user's default browser to 10 seconds for running on slower computers.  It would be nice to figure out a way to detect that jetty has completed startup instead of guessing.