OrderedMap needs to be thread-safe.
To make it thread-safe, instead of inheriting from java.util.HashMap,
inherit from java.util.concurrent.ConcurrentHashMap. There are two
things to note.

1. The semantics of ConcurrentHashMap.get() are different from the
semantics of HashMap.get() when the provided key is null. This change
duplicates the semantics of HashMap as other parts of Mockey rely on
that.

2. This doesn't really make OrderedMap thread-safe. It allows
concurrent modifications to happen, but that assumes that the patterns
in which OrderedMap are used are in themselves thread-safe.